Evolution: cron Cleanoldagent update the status of the item instead of removing the agent. Adding an extra config option with either delete or change the agent status

+++ inc/agent.class.php
@@ -1217,7 +1217,7 @@ class PluginFusioninventoryAgent extends CommonDBTM {
    * @return bool cron is ok or not
    *
    **/
-   static function cronCleanoldagents() {
+   static function cronCleanoldagents($task=NULL) {
       global $DB;
 
       $pfConfig = new PluginFusioninventoryConfig();
@@ -1225,18 +1225,37 @@ class PluginFusioninventoryAgent extends CommonDBTM {
 
       $retentiontime = $pfConfig->getValue('agents_old_days');
       if ($retentiontime == 0) {
-         return TRUE;
+         return 0;
       }
       $sql = "SELECT * FROM `glpi_plugin_fusioninventory_agents`
-                WHERE `last_contact` < date_add(now(), interval -".$retentiontime." day)";
-      $result=$DB->query($sql);
+                   WHERE `last_contact` < date_add(now(), interval -".$retentiontime." day)";
+      $result = $DB->query($sql);
+
       if ($result) {
-         while ($data=$DB->fetch_array($result)) {
-            $pfAgent->delete($data);
+         $cron_status = 0;
+         $action = $pfConfig->getValue('agents_action');
+         if ($action == PluginFusioninventoryConfig::ACTION_CLEAN) {
+            //delete agents
+            while ($data = $DB->fetch_array($result)) {
+               $pfAgent->delete($data);
+                $task->addVolume(1);
+                $cron_status = 1;
+            }
+         } else {
+            //change status of agents
+            while ($data = $DB->fetch_array($result)) {
+               $computer = new Computer();
+               if($computer->getFromDB($data['computers_id'])){
+                  $computer->update(array('id' => $data['computers_id'], 'states_id' => $pfConfig->getValue('status')));
+                   $task->addVolume(1);
+                   $cron_status = 1;
+               }
+            }
          }
       }
-      return TRUE;
+      return $cron_status;
    }
+
 }
 
 ?>
